Redefine whitespace as "tab or space" for filenames
After considering the pros and cons, this seems like the best way to get
around the issue of false-positive space character detection with the \s
character group.

Decoding the filenames as utf8 would be a lossy conversion, and would
require too much "understanding" of the filenames and checklist file
contents IMO. It would make it impossible to express some filenames in
the checklist file, and it would introduce nasty bugs with globbing
those unsupported filenames.

With that in mind I've again opted to keep everything in "ascii plus
other stuff I don't need to understand" encoding.

Specifically, one file causing problems ended in 情.
Perl will consider this character to end in whitespace if all of the
following are true:
- This character is treated as a string of utf-8 bytes
- Perl has "use 5.011" or above active